Home page
Implements a home page that displays user profiles using a javascript function and allows the user to like/dislike the profile.

These profiles come from a list returned from profilearray.py.

Right now, the logic doesn't exist for us to loop through the actual userids in the database (see TODO in profilearray.py) so the array must be manually changed to reflect what user ids are in your database.

There is a function call when a profile is liked/disliked that currently does nothing. Need to implement some kind of logic there.

Adds home to navbar